+++ cloud-spanner-r2dbc/src/main/java/com/google/cloud/spanner/r2dbc/v2/ClientLibraryCodec.java
@@ -9,34 +9,38 @@ import com.google.cloud.spanner.Struct;
 import com.google.cloud.spanner.Type;
 
 public class ClientLibraryCodec {
-	private static Map<Type, BiFunction<Struct, Integer, Object>> getters = new HashMap<>();
-	static {
-		getters.put(Type.int64(), AbstractStructReader::getLong);
-		getters.put(Type.array(Type.int64()), AbstractStructReader::getLongList);
+	private static Map<Type, BiFunction<Struct, Integer, Object>> codecsMap = createCodecsMap();
 
-		getters.put(Type.float64(), AbstractStructReader::getDouble);
-		getters.put(Type.array(Type.float64()), AbstractStructReader::getDoubleList);
+	private static Map<Type, BiFunction<Struct, Integer, Object>> createCodecsMap() {
+		Map<Type, BiFunction<Struct, Integer, Object>> codecs = new HashMap<>();
+		codecs.put(Type.int64(), AbstractStructReader::getLong);
+		codecs.put(Type.array(Type.int64()), AbstractStructReader::getLongList);
 
-		getters.put(Type.bool(), AbstractStructReader::getBoolean);
-		getters.put(Type.array(Type.bool()), AbstractStructReader::getBooleanList);
+		codecs.put(Type.float64(), AbstractStructReader::getDouble);
+		codecs.put(Type.array(Type.float64()), AbstractStructReader::getDoubleList);
 
-		getters.put(Type.bytes(), AbstractStructReader::getBytes);
-		getters.put(Type.array(Type.bytes()), AbstractStructReader::getBytesList);
+		codecs.put(Type.bool(), AbstractStructReader::getBoolean);
+		codecs.put(Type.array(Type.bool()), AbstractStructReader::getBooleanList);
 
-		getters.put(Type.date(), AbstractStructReader::getDate);
-		getters.put(Type.array(Type.date()), AbstractStructReader::getDateList);
+		codecs.put(Type.bytes(), AbstractStructReader::getBytes);
+		codecs.put(Type.array(Type.bytes()), AbstractStructReader::getBytesList);
 
-		getters.put(Type.string(), AbstractStructReader::getString);
-		getters.put(Type.array(Type.string()), AbstractStructReader::getStringList);
+		codecs.put(Type.date(), AbstractStructReader::getDate);
+		codecs.put(Type.array(Type.date()), AbstractStructReader::getDateList);
 
-		getters.put(Type.timestamp(), AbstractStructReader::getTimestamp);
-		getters.put(Type.array(Type.timestamp()), AbstractStructReader::getTimestampList);
+		codecs.put(Type.string(), AbstractStructReader::getString);
+		codecs.put(Type.array(Type.string()), AbstractStructReader::getStringList);
+
+		codecs.put(Type.timestamp(), AbstractStructReader::getTimestamp);
+		codecs.put(Type.array(Type.timestamp()), AbstractStructReader::getTimestampList);
+
+		return codecs;
 	}
 
 	public static  <T> T decode(Struct struct, int index, Class<T> type) {
 		Object value = struct.isNull(index)
 				? null
-				: getters.get(struct.getColumnType(index)).apply(struct, index);
+				: codecsMap.get(struct.getColumnType(index)).apply(struct, index);
 		return (T) value;
 	}
 }
